fix: add createHandler file
