+++ GestureHandler.js
@@ -1,6 +1,5 @@
 import React from 'react';
 import {
-  findNodeHandle,
   Animated,
   ScrollView,
   Slider,
@@ -14,22 +13,16 @@ import {
   FlatList,
   Platform,
   processColor,
-  Touchable,
 } from 'react-native';
-import deepEqual from 'fbjs/lib/areEqual';
 import PropTypes from 'prop-types';
 
+import createHandler from './createHandler';
 import GestureHandlerButton from './GestureHandlerButton';
-import GestureHandlerModule from './GestureHandlerModule';
-import GestureHandlerRenderer from './GestureHandlerRenderer';
 import gestureHandlerRootHOC from './gestureHandlerRootHOC';
 
 import Directions from './Directions';
 import State from './State';
 import PlatformConstants from './PlatformConstants';
-
-let handlerTag = 1;
-const handlerIDToTag = {};
 
 const GestureHandlerPropTypes = {
   id: PropTypes.string,
@@ -72,305 +65,15 @@ const GestureHandlerPropTypes = {
   onEnded: PropTypes.func,
 };
 
-const stateToPropMappings = {
-  [State.BEGAN]: 'onBegan',
-  [State.FAILED]: 'onFailed',
-  [State.CANCELLED]: 'onCancelled',
-  [State.ACTIVE]: 'onActivated',
-  [State.END]: 'onEnded',
-};
-
-function isConfigParam(param, name) {
-  return (
-    param !== undefined &&
-    typeof param !== 'function' &&
-    (typeof param !== 'object' || !('__isNative' in param)) &&
-    name !== 'onHandlerStateChange' &&
-    name !== 'onGestureEvent'
-  );
-}
-
-function transformIntoHandlerTags(handlerIDs) {
-  if (!Array.isArray(handlerIDs)) {
-    handlerIDs = [handlerIDs];
-  }
-
-  // converts handler string IDs into their numeric tags
-  return handlerIDs
-    .map(
-      handlerID =>
-        handlerIDToTag[handlerID] ||
-        (handlerID.current && handlerID.current._handlerTag) ||
-        -1
-    )
-    .filter(handlerTag => handlerTag > 0);
-}
-
-function filterConfig(props, validProps, defaults = {}) {
-  const res = { ...defaults };
-  Object.keys(validProps).forEach(key => {
-    const value = props[key];
-    if (isConfigParam(value, key)) {
-      let value = props[key];
-      if (key === 'simultaneousHandlers' || key === 'waitFor') {
-        value = transformIntoHandlerTags(props[key]);
-      } else if (key === 'hitSlop') {
-        if (typeof value !== 'object') {
-          value = { top: value, left: value, bottom: value, right: value };
-        }
-      }
-      res[key] = value;
-    }
-  });
-  return res;
-}
-
-function hasUnresolvedRefs(props) {
-  const extract = refs => {
-    if (!Array.isArray(refs)) {
-      return refs && refs.current === null;
-    }
-    return refs.some(r => r && r.current === null);
-  };
-  return extract(props['simultaneousHandlers']) || extract(props['waitFor']);
-}
-
-function createHandler(
-  handlerName,
-  propTypes = null,
-  config = {},
-  transformProps,
-  customNativeProps = {}
-) {
-  class Handler extends React.Component {
-    static displayName = handlerName;
-
-    static propTypes = {
-      ...GestureHandlerPropTypes,
-      ...propTypes,
-    };
-
-    constructor(props) {
-      super(props);
-      this._handlerTag = handlerTag++;
-      this._config = {};
-      if (props.id) {
-        if (handlerIDToTag[props.id] !== undefined) {
-          throw new Error(`Handler with ID "${props.id}" already registered`);
-        }
-        handlerIDToTag[props.id] = this._handlerTag;
-      }
-    }
-
-    _onGestureHandlerEvent = event => {
-      if (event.nativeEvent.handlerTag === this._handlerTag) {
-        this.props.onGestureEvent && this.props.onGestureEvent(event);
-      } else {
-        this.props.onGestureHandlerEvent &&
-          this.props.onGestureHandlerEvent(event);
-      }
-    };
-
-    _onGestureHandlerStateChange = event => {
-      if (event.nativeEvent.handlerTag === this._handlerTag) {
-        this.props.onHandlerStateChange &&
-          this.props.onHandlerStateChange(event);
-
-        const stateEventName = stateToPropMappings[event.nativeEvent.state];
-        if (typeof this.props[stateEventName] === 'function') {
-          this.props[stateEventName](event);
-        }
-      } else {
-        this.props.onGestureHandlerStateChange &&
-          this.props.onGestureHandlerStateChange(event);
-      }
-    };
-
-    _refHandler = node => {
-      this._viewNode = node;
-
-      const child = React.Children.only(this.props.children);
-      const { ref } = child;
-      if (ref !== null) {
-        if (typeof ref === 'function') {
-          ref(node);
-        } else {
-          ref.current = node;
-        }
-      }
-    };
-
-    _createGestureHandler = newConfig => {
-      this._config = newConfig;
-
-      if (!GestureHandlerModule.createGestureHandler) {
-        return;
-      }
-
-      GestureHandlerModule.createGestureHandler(
-        handlerName,
-        this._handlerTag,
-        newConfig
-      );
-    };
-
-    _attachGestureHandler = newViewTag => {
-      this._viewTag = newViewTag;
-
-      if (!GestureHandlerModule.attachGestureHandler) {
-        return;
-      }
-
-      GestureHandlerModule.attachGestureHandler(this._handlerTag, newViewTag);
-    };
-
-    _updateGestureHandler = newConfig => {
-      this._config = newConfig;
-
-      if (!GestureHandlerModule.updateGestureHandler) {
-        return;
-      }
-
-      GestureHandlerModule.updateGestureHandler(this._handlerTag, newConfig);
-    };
-
-    _dropGestureHandler = () => {
-      if (!GestureHandlerModule.dropGestureHandler) {
-        return;
-      }
-
-      GestureHandlerModule.dropGestureHandler(this._handlerTag);
-    };
-
-    componentWillUnmount() {
-      if (this._updateEnqueued) {
-        clearImmediate(this._updateEnqueued);
-      }
-      if (this.props.id) {
-        delete handlerIDToTag[this.props.id];
-      }
-    }
-
-    componentDidMount() {
-      if (hasUnresolvedRefs(this.props)) {
-        // If there are unresolved refs (e.g. ".current" has not yet been set)
-        // passed as `simultaneousHandlers` or `waitFor`, we enqueue a call to
-        // _update method that will try to update native handler props using
-        // setImmediate. This makes it so _update function gets called after all
-        // react components are mounted and we expect the missing ref object to
-        // be resolved by then.
-        this._updateEnqueued = setImmediate(() => {
-          this._updateEnqueued = null;
-          this._update();
-        });
-      }
-
-      this._createGestureHandler(
-        filterConfig(
-          transformProps ? transformProps(this.props) : this.props,
-          { ...this.constructor.propTypes, ...customNativeProps },
-          config
-        )
-      );
-      this._attachGestureHandler(findNodeHandle(this._viewNode));
-    }
-
-    componentDidUpdate() {
-      const viewTag = findNodeHandle(this._viewNode);
-      if (this._viewTag !== viewTag) {
-        this._attachGestureHandler(viewTag);
-      }
-      this._update();
-    }
-
-    _update() {
-      const newConfig = filterConfig(
-        transformProps ? transformProps(this.props) : this.props,
-        { ...this.constructor.propTypes, ...customNativeProps },
-        config
-      );
-      if (!deepEqual(this._config, newConfig)) {
-        this._updateGestureHandler(newConfig);
-      }
-    }
-
-    setNativeProps(updates) {
-      const mergedProps = { ...this.props, ...updates };
-      const newConfig = filterConfig(
-        transformProps ? transformProps(mergedProps) : mergedProps,
-        { ...this.constructor.propTypes, ...customNativeProps },
-        config
-      );
-      this._updateGestureHandler(newConfig);
-    }
-
-    render() {
-      let gestureEventHandler = this._onGestureHandlerEvent;
-      const { onGestureEvent, onGestureHandlerEvent } = this.props;
-      if (onGestureEvent && typeof onGestureEvent !== 'function') {
-        // If it's not a method it should be an native Animated.event
-        // object. We set it directly as the handler for the view
-        // In this case nested handlers are not going to be supported
-        if (onGestureHandlerEvent) {
-          throw new Error(
-            'Nesting touch handlers with native animated driver is not supported yet'
-          );
-        }
-        gestureEventHandler = this.props.onGestureEvent;
-      } else {
-        if (
-          onGestureHandlerEvent &&
-          typeof onGestureHandlerEvent !== 'function'
-        ) {
-          throw new Error(
-            'Nesting touch handlers with native animated driver is not supported yet'
-          );
-        }
-      }
-
-      let gestureStateEventHandler = this._onGestureHandlerStateChange;
-      const { onHandlerStateChange, onGestureHandlerStateChange } = this.props;
-      if (onHandlerStateChange && typeof onHandlerStateChange !== 'function') {
-        // If it's not a method it should be an native Animated.event
-        // object. We set it directly as the handler for the view
-        // In this case nested handlers are not going to be supported
-        if (onGestureHandlerStateChange) {
-          throw new Error(
-            'Nesting touch handlers with native animated driver is not supported yet'
-          );
-        }
-        gestureStateEventHandler = this.props.onHandlerStateChange;
-      } else {
-        if (
-          onGestureHandlerStateChange &&
-          typeof onGestureHandlerStateChange !== 'function'
-        ) {
-          throw new Error(
-            'Nesting touch handlers with native animated driver is not supported yet'
-          );
-        }
-      }
-
-      return GestureHandlerRenderer.render({
-        ...this.props,
-        handlerName,
-        ref: this._refHandler,
-        collapsable: false,
-        onGestureHandlerEvent: gestureEventHandler,
-        onGestureHandlerStateChange: gestureStateEventHandler,
-      });
-    }
-  }
-  return Handler;
-}
-
 const NativeViewGestureHandler = createHandler('NativeViewGestureHandler', {
+  ...GestureHandlerPropTypes,
   shouldActivateOnStart: PropTypes.bool,
   disallowInterruption: PropTypes.bool,
 });
 const TapGestureHandler = createHandler(
   'TapGestureHandler',
   {
+    ...GestureHandlerPropTypes,
     maxDurationMs: PropTypes.number,
     maxDelayMs: PropTypes.number,
     numberOfTaps: PropTypes.number,
@@ -385,6 +88,7 @@ const TapGestureHandler = createHandler(
 const FlingGestureHandler = createHandler(
   'FlingGestureHandler',
   {
+    ...GestureHandlerPropTypes,
     numberOfPointers: PropTypes.number,
     direction: PropTypes.number,
   },
@@ -406,6 +110,7 @@ const ForceTouchGestureHandler = PlatformConstants.forceTouchAvailable
   ? createHandler(
       'ForceTouchGestureHandler',
       {
+        ...GestureHandlerPropTypes,
         minForce: PropTypes.number,
         maxForce: PropTypes.number,
         feedbackOnActivation: PropTypes.bool,
@@ -420,6 +125,7 @@ ForceTouchGestureHandler.forceTouchAvailable =
 const LongPressGestureHandler = createHandler(
   'LongPressGestureHandler',
   {
+    ...GestureHandlerPropTypes,
     minDurationMs: PropTypes.number,
     maxDist: PropTypes.number,
   },
@@ -586,6 +292,7 @@ function managePanProps(props) {
 const PanGestureHandler = createHandler(
   'PanGestureHandler',
   {
+    ...GestureHandlerPropTypes,
     activeOffsetY: PropTypes.oneOfType([
       PropTypes.number,
       PropTypes.arrayOf(PropTypes.number),
@@ -623,8 +330,16 @@ const PanGestureHandler = createHandler(
     failOffsetXEnd: true,
   }
 );
-const PinchGestureHandler = createHandler('PinchGestureHandler', {}, {});
-const RotationGestureHandler = createHandler('RotationGestureHandler', {}, {});
+const PinchGestureHandler = createHandler(
+  'PinchGestureHandler',
+  GestureHandlerPropTypes,
+  {}
+);
+const RotationGestureHandler = createHandler(
+  'RotationGestureHandler',
+  GestureHandlerPropTypes,
+  {}
+);
 
 const NATIVE_WRAPPER_BIND_BLACKLIST = new Set(['replaceState', 'isMounted']);
 const NATIVE_WRAPPER_PROPS_FILTER = {
