+++ package.json
@@ -17,16 +17,14 @@
     "android/lib/src/main/java/",
     "ios/",
     "touchables/**/*.js",
+    "createHandler.js",
+    "createHandler.web.js",
     "Directions.js",
     "Directions.web.js",
     "DrawerLayout.js",
     "GestureHandler.js",
     "GestureHandlerButton.js",
     "GestureHandlerButton.web.js",
-    "GestureHandlerModule.js",
-    "GestureHandlerModule.web.js",
-    "GestureHandlerRenderer.js",
-    "GestureHandlerRenderer.web.js",
     "gestureHandlerRootHOC.android.js",
     "gestureHandlerRootHOC.ios.js",
     "gestureHandlerRootHOC.js",
